requirements.txt
# Hyperliquid API Client
hyperliquid-python-sdk>=0.4.0

# Environment Variables
python-dotenv>=1.0.0

# Database (SQLite is built-in to Python)

# Utilities
pandas>=2.0.0  # For PnL dashboard
tabulate>=0.9.0  # For CLI table formatting
